Int to long commmit

+++ Bot.Business/Abstractions/ServicesAbstractions/IDeleteMessageService.cs
@@ -9,7 +9,7 @@ namespace Bot.Business.Abstractions.ServicesAbstractions
 {
     public interface IDeleteMessageService
     {
-        void DeleteMessage(int messageId);
-        List<Message> GetDeletedMessages(int? userId, int chatId);
+        void DeleteMessage(long messageId);
+        List<Message> GetDeletedMessages(long? userId, long chatId);
     }
 }
