+++ Bot.Data/Models/Message.cs
@@ -7,15 +7,15 @@ namespace Bot.Data.Models
     {
         [BsonId]
         [BsonRepresentation(BsonType.ObjectId)]
-        public int Id { get; set; }
-        public int UserId { get; set; }
-        public int ChatId { get; set; }
-        public int Date { get; set; }
+        public long Id { get; set; }
+        public long UserId { get; set; }
+        public long ChatId { get; set; }
+        public long Date { get; set; }
         public string Text { get; set; }
         public bool ContainsAudio { get; set; }
-        public int Photo { get; set; }
-        public int Sticker { get; set; }
-        public int Animation { get; set; }
+        public long Photo { get; set; }
+        public long Sticker { get; set; }
+        public long Animation { get; set; }
         public string Caption { get; set; }
     }
 }
