+++ Bot.Data/Models/Mute.cs
@@ -7,11 +7,11 @@ namespace Bot.Data.Models
     {
         [BsonId]
         [BsonRepresentation(BsonType.ObjectId)]
-        public int Id { get; set; }
-        public int UserId { get; set; }
-        public int ChatId { get; set; }
+        public long Id { get; set; }
+        public long UserId { get; set; }
+        public long ChatId { get; set; }
         public int MuteDuration { get; set; }
         public int MuteTime { get; set; }
-        public int MessageId { get; set; }
+        public long MessageId { get; set; }
     }
 }
